User auth/redux/37/save data (#66)
* Created routing and components for signup and home pages

* feat: add auth services

* Added bootstrap to index

* Feat: Button and input first styles done

* feat: user auth middleware

* feat: firebase implementation backend done

* Conditional classes done

* feat: getuser endpoint created

* Feat: styles done

* Feat: App test remove

* Created main firebase functions in /services

* refactor: set api/ as start of endpoint

* Finished main implementation

* feat: created endpoint to update a field of the user

* Changed routes

* feat: create account with firebase

* First formulary structure done

* refactor: getUser with firebase implemented

* feat: created update account with firebase middleware

* refactor: removed id from routes on update account

* Feat: profilePicture updated

* Feat: Pre-cleaning

* Feat: Update done

* First updates done

* Feat: Functionality done

* Feat: Creating Select component

* Feat: Select component done

* Feat: Select component bug fixed

* Feat: User account showing and editing done Closes User info visualization #12  Closes User info edition #13

* feat: signup and layout WIP

* feat: style

* Installed react-redux

* Added redux with userState and auth on start

* Feat: user redux state on page load set

* feat: layout WIP

* Added deleted firebase files

* Feat: sign out and header navigation

* Feat: PrivateRoute component

* feat: onlyPublicRoutes component created and implemented in App

* bug: removed console log

* cleaned code

* Added email verification on sign up

* Added account links and password redirect
