+++ packages/client/src/components/OnlyPublicRoute/index.js
@@ -1,0 +1,1 @@
+export { default } from "./OnlyPublicRoute";
