+++ packages/client/src/components/PrivateRoute/index.js
@@ -1,0 +1,1 @@
+export { default } from "./PrivateRoute";
