+++ packages/client/src/pages/Public/SignUp/SignUp.scss
@@ -1,0 +1,9 @@
+@use "../../../sass/components/colors" as c;
+
+.fnt-primary{
+    color: c.$primaryColor;
+}
+
+.fnt-secondary{
+    color: c.$secondaryColor;
+}
