+++ packages/client/src/redux/user/state.js
@@ -1,0 +1,11 @@
+const initialState = {
+  email: "",
+  token: "",
+  firstName: "",
+  lastName: "",
+  profilePicture: "",
+  firebaseId: "",
+  isLogged: false,
+};
+
+export default initialState;
