+++ packages/client/src/redux/user/types.js
@@ -1,0 +1,2 @@
+export const SET_LOGIN = "SET_LOGIN";
+export const SET_LOGOUT = "SET_LOGOUT";
