+++ packages/server/package.json
@@ -41,8 +41,5 @@
     "mongoose": "^5.13.7",
     "morgan": "^1.10.0",
     "validator": "^13.6.0"
-  },
-  "engines": {
-    "node": ">=14.0.0"
   }
 }
